chore: removed preview.js for prefered .jsx extension
